Read me.md
A genetic algorithm implementation of the classic N-Queens problem, written in Python. The N-Queens problem is a well-known problem in the field of computer science and artificial intelligence, which involves placing N queens on an NxN chessboard in such a way that no two queens threaten each other. This problem can be extended to other similar problems like n-rooks, n-bishops, n-knights and so on.